requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==2.0.0
Markdown==3.3.6
Pillow==9.0.1
Pygments==2.11.2
sqlparse==0.4.2
tzdata==2022.1
